backend/requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.52.1
fsspec==2024.5.0
ftfy==6.2.0
idna==3.7
imageio==2.34.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.22
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.22
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
